requirements.txt
# FastAPI and web framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Graph databases
falkordb
graphiti-core

# Database drivers
mysql-connector-python
pymysql
JPype1
jaydebeapi
pyodbc
sqlalchemy
pymongo

# LLM Integration
openai

# HTTP and API clients
requests
httpx

# Utilities
python-dotenv
python-multipart
filelock
croniter
python-dateutil
tenacity
orjson
jinja2

# Logging and monitoring
structlog
prometheus-client
psutil

# Development dependencies (optional)
pytest
pytest-asyncio
black
flake8
mypy

